+++ crates/genesis/Cargo.toml
@@ -19,8 +19,8 @@ alloy-signer.workspace = true
 
 # serde
 serde.workspace = true
-serde_json.workspace = true
 k256.workspace = true
 
-
+[dev-dependencies]
+serde_json.workspace = true
