+++ crates/genesis/README.md
@@ -1,0 +1,13 @@
+# alloy-genesis
+
+The genesis file in an Ethereum blockchain is a JSON-formatted file used to define the initial state of the blockchain at the time of its creation. alloy-genesis provides tools and structures to create, manipulate, and interpret these genesis files, ensuring seamless integration and configuration for Ethereum-compatible networks.
+
+## Features
+
+- `Genesis Block Configuration`- Facilitates the creation and definition of genesis blocks, including parameters such as nonce, timestamp, gas limits, and initial account states.
+- `Custom Account Initialization`- Allows for the specification of custom accounts with predefined balances, nonces, and contract codes in the genesis block.
+- `Network Compatibility`- Designed to work with Ethereum-like networks, ensuring compatibility with various Ethereum standards.
+
+## Usage
+
+This crate is primarily used internally within the Alloy project and is not intended for direct use. However, it can be utilized by blockchain developers or maintainers for setting up new Ethereum-like networks or for testing purposes.
